feat: Add option for pre-release updates
Introduces a new setting to allow users to opt-in to receiving pre-release (beta) versions of the application. This includes updating the settings type, default settings, storage service interface, and UI elements (README, functional manual, technical manual) to accommodate this new feature. The `useSettings` hook is updated to notify the Electron main process when this setting changes.